mcsampler.py: yet another commit to deal with old float128/float64 typing

+++ MonteCarloMarginalizeCode/Code/RIFT/integrators/mcsampler.py
@@ -190,12 +190,12 @@ class MCSampler(object):
         def dP_cdf(p, x):
             if x > self.rlim[param] or x < self.llim[param]:
                 return 0
-            return self.pdf[param](x)
-        x_i = numpy.linspace(self.llim[param], self.rlim[param], 1000)
+            return numpy.float64(self.pdf[param](x))
+        x_i = numpy.linspace(self.llim[param], self.rlim[param], 1000,dtype=numpy.float64)
         # Integrator needs to have a step size which doesn't step over the
         # probability mass
         # TODO: Determine h_max.
-        cdf = integrate.odeint(dP_cdf, [0], x_i, hmax=0.01*(self.rlim[param]-self.llim[param])).T[0]
+        cdf = integrate.odeint(dP_cdf, [0], x_i, hmax=0.01*numpy.float64(self.rlim[param]-self.llim[param])).T[0]
         if cdf[-1] != 1.0: # original pdf wasn't normalized
             self._pdf_norm[param] = cdf[-1]
             cdf /= cdf[-1]
